change total rating type, rework action filter

+++ GameStore/BL/Utilities/PagedList.cs
@@ -10,27 +10,27 @@ namespace GameStore.BL.Utilities
         {
         }
 
-        public PagedList(IEnumerable<T> items, int count, int pageNumber, int pageSize)
+        public PagedList(IEnumerable<T> items, uint count, uint pageNumber, uint pageSize)
         {
             TotalCount = count;
             PageSize = pageSize;
             CurrentPage = pageNumber;
-            TotalPages = (int)Math.Ceiling(count / (double)pageSize);
+            TotalPages = (uint)Math.Ceiling(count / (double)pageSize);
             AddRange(items);
         }
 
-        public int CurrentPage { get; }
-        public int TotalPages { get; }
-        public int PageSize { get; }
-        public int TotalCount { get; }
+        public uint CurrentPage { get; }
+        public uint TotalPages { get; }
+        public uint PageSize { get; }
+        public uint TotalCount { get; }
         public bool HasPrevious => CurrentPage > 1;
         public bool HasNext => CurrentPage < TotalPages;
 
-        public static PagedList<T> ToPagedList(IQueryable<T> source, int pageNumber, int pageSize)
+        public static PagedList<T> ToPagedList(IQueryable<T> source, uint pageNumber, uint pageSize)
         {
             var count = source.Count();
-            var items = source.Skip((pageNumber - 1) * pageSize).Take(pageSize).ToList();
-            return new(items, count, pageNumber, pageSize);
+            var items = source.Skip((int)((pageNumber - 1) * pageSize)).Take((int)pageSize).ToList();
+            return new(items, (uint)count, pageNumber, pageSize);
         }
     }
 }
